Fix steps to build test container for Ansible
This follows the best practices recommended by:
https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/#/run

In particular, it addresses caching issues that can occur for raw "RUN
apt-get update" commands.

This is a little crummy at the moment due to the need to install tools
to add a PPA repository. It may be worth investigating an alternative
base (i.e. Debian).